requirements.txt
openai-whisper
tts
ffmpeg-python
requests
fastapi
uvicorn
twilio
python-dotenv 